Using Client API on Home page
Now using api to get example JSON information. The goals are the same regardless of which student ID you choose, but the moral of the story is we are actually calling async functions to get this info. When the client API is connected to the db API we will get all sorts of data